Realistic fall/jump physics, fixed jump in air glitch, improved the player\enemies code by combining their movement into the separate method of the entity class

+++ GAME1017_Template_W01/PlatformPlayer.h
@@ -6,9 +6,9 @@
 #include "Fireball.h"
 #include "Entity.h"
 
-const int MELEECOOLDOWN = 1.0;
-const int MAGICCOOLDOWN = 1.0;
-const int FIREBALLCOST = 10.0;
+const double MELEECOOLDOWN = 1.0;
+const double MAGICCOOLDOWN = 1.0;
+const double FIREBALLCOST = 10.0;
 
 class PlatformPlayer : public Entity
 {
